Refactor Shapeviewer - file was becoming oversized
